+++ src/ios/HealthKit.m
@@ -803,7 +803,10 @@ static NSString *const HKPluginKeyUUID = @"UUID";
     //    workoutPredicate = [HKQuery predicateForWorkoutsWithWorkoutActivityType:HKWorkoutActivityTypeCycling];
     //  }
 
-    BOOL *includeCalsAndDist = (args[@"includeCalsAndDist"] != nil && [args[@"includeCalsAndDist"] boolValue]);
+    BOOL *includeCalories = (args[@"includeCalories"] != nil && [args[@"includeCalories"] boolValue]);
+    BOOL *includeDistance = (args[@"includeDistance"] != nil && [args[@"includeDistance"] boolValue]);
+
+    
 
     NSSet *types = [NSSet setWithObjects:[HKWorkoutType workoutType], nil];
     [[HealthKit sharedHealthStore] requestAuthorizationToShareTypes:nil readTypes:types completion:^(BOOL success, NSError *error) {
@@ -835,40 +838,30 @@ static NSString *const HKPluginKeyUUID = @"UUID";
                             source = workout.source;
                         }
                         NSMutableDictionary *entry;
+                        entry = [
+                                @{
+                                        @"duration": @(workout.duration),
+                                        HKPluginKeyStartDate: [HealthKit stringFromDate:workout.startDate],
+                                        HKPluginKeyEndDate: [HealthKit stringFromDate:workout.endDate],
+                                        HKPluginKeySourceBundleId: source.bundleIdentifier,
+                                        HKPluginKeySourceName: source.name,
+                                        @"activityType": workoutActivity,
+                                        @"UUID": [workout.UUID UUIDString]
+                                } mutableCopy
+                            ];
 
-                        if(includeCalsAndDist != nil && includeCalsAndDist) {
-                            double meters = [workout.totalDistance doubleValueForUnit:[HKUnit meterUnit]];
-                            NSString *metersString = [NSString stringWithFormat:@"%ld", (long) meters];
-
+                        if(includeCalories != nil && includeCalories) {
                             // Parse totalEnergyBurned in kilocalories
                             double cals = [workout.totalEnergyBurned doubleValueForUnit:[HKUnit kilocalorieUnit]];
                             NSString *calories = [[NSNumber numberWithDouble:cals] stringValue];
 
-                            entry = [
-                                @{
-                                        @"duration": @(workout.duration),
-                                        HKPluginKeyStartDate: [HealthKit stringFromDate:workout.startDate],
-                                        HKPluginKeyEndDate: [HealthKit stringFromDate:workout.endDate],
-                                        @"distance": metersString,
-                                        @"energy": calories,
-                                        HKPluginKeySourceBundleId: source.bundleIdentifier,
-                                        HKPluginKeySourceName: source.name,
-                                        @"activityType": workoutActivity,
-                                        @"UUID": [workout.UUID UUIDString]
-                                } mutableCopy
-                            ];
-                        }  else {
-                            entry = [
-                                @{
-                                        @"duration": @(workout.duration),
-                                        HKPluginKeyStartDate: [HealthKit stringFromDate:workout.startDate],
-                                        HKPluginKeyEndDate: [HealthKit stringFromDate:workout.endDate],
-                                        HKPluginKeySourceBundleId: source.bundleIdentifier,
-                                        HKPluginKeySourceName: source.name,
-                                        @"activityType": workoutActivity,
-                                        @"UUID": [workout.UUID UUIDString]
-                                } mutableCopy
-                            ];
+                            entry[@"energy"] = calories;
+                        }
+                        if(includeDistance != nil && includeDistance) {
+                            double meters = [workout.totalDistance doubleValueForUnit:[HKUnit meterUnit]];
+                            NSString *metersString = [NSString stringWithFormat:@"%ld", (long) meters];
+
+                            entry[@"distance"] = metersString;
                         }
 
                         [finalResults addObject:entry];